requirements.txt
fastapi[standard]>=0.115.0
uvicorn[standard]>=0.30.0
jinja2>=3.1.0
python-dotenv>=1.0.0
pydantic-settings>=2.0.0
anthropic>=0.40.0
httpx>=0.27.0
browserbase>=1.0.0
playwright>=1.48.0
notion-client>=2.2.0
